Extract shared source/dest methods to Registry::Fileable
